# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSK Smart Learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile and HSK goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding with placement result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        currentLevel:
          type: integer
        targetLevel:
          type: integer
        onboardingCompleted:
          type: boolean
        totalXp:
          type: integer
        dailyGoalMinutes:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - currentLevel
        - targetLevel
        - onboardingCompleted
        - totalXp
        - dailyGoalMinutes
        - createdAt
    UserProfileInput:
      type: object
      properties:
        name:
          type: string
        targetLevel:
          type: integer
        dailyGoalMinutes:
          type: integer
    OnboardingInput:
      type: object
      properties:
        name:
          type: string
        targetLevel:
          type: integer
        placementScore:
          type: integer
        recommendedLevel:
          type: integer
        dailyGoalMinutes:
          type: integer
      required:
        - name
        - targetLevel
        - recommendedLevel
...